requirements.txt
# V5-Notebook-Helper Dependencies

# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# OpenAI API
openai>=1.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2

# CLI
typer>=0.9.0
rich>=13.0.0

# Image Processing
Pillow>=10.0.0

# Data Processing
pyyaml>=6.0.0
python-multipart>=0.0.6

# Utilities
httpx>=0.25.0
